+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add `strokeAnimation` option to Svg component
 - Add `stroke` option to Parallax component
+- Add `dataSrc` as primary option to Image component
 - Parallax component supports more units
 
 ### Changed